Change tests folder & Add app test
