+++ src/__tests__/button.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import Button from './Button';
+import Button from '../components/Button';
 
 describe('Button', () => {
   test('Link renders correctly', () => {
